+++ packages/master/package.json
@@ -35,17 +35,13 @@
     "yargs": "^14.2.0"
   },
   "devDependencies": {
-    "@ant-design/icons": "^4.2.2",
     "@babel/core": "^7.11.6",
     "@babel/preset-react": "^7.10.4",
-    "antd": "^4.6.5",
+    "@clusterio/web_ui": "^2.0.0-alpha.1",
     "babel-loader": "^8.1.0",
-    "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^4.3.0",
-    "events": "^3.2.0",
     "fast-deep-equal": "^3.1.3",
     "null-loader": "^4.0.0",
-    "path-browserify": "^1.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
