+++ packages/web_ui/src/pages/slaves/slavesTable.jsx
