+++ plugins/global_chat/package.json
@@ -19,12 +19,9 @@
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/preset-react": "^7.10.4",
-    "@clusterio/master": "^2.0.0-alpha.1",
+    "@clusterio/web_ui": "^2.0.0-alpha.1",
     "babel-loader": "^8.1.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "events": "^3.2.0",
     "null-loader": "^4.0.0",
-    "path-browserify": "^1.0.1",
     "webpack-cli": "^3.3.12",
     "webpack-merge": "^5.2.0"
   },
